be/django/requirements.txt
django
django-filter
django-guardian
djangorestframework
djangorestframework-jwt
django-rest-auth
django-cors-headers

aldjemy
werkzeug
psycopg2
